+++ docs/changelog.rst
@@ -1,6 +1,11 @@
 Changelog
 =========
 
+development
+-----------
+
+* Changed ``defopt.run`` to return the value from the called function
+
 1.0.1 (2016-02-14)
 ------------------